Fix expired status mapping, enable decimal weight input, add slide-over detail view for P3K/APAR/Hydrant, enhance asset info mapping, and resolve asset type retrieval issue

+++ database/migrations/2026_02_05_072922_create_apars_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->foreignId("apar_type_id")->constrained("apar_types");
             $table->foreignId('room_id')->constrained('rooms');
             $table->enum('status', ['safe', 'warning', 'critical'])->default('safe');
-            $table->integer('weight');
+            $table->decimal('weight');
             $table->date('last_refilled_at')->nullable();
             $table->date('expired_at');
             $table->json('location_data')->nullable();
